Bug fix get article and get user done

+++ application/controllers/admin/article.php
@@ -38,19 +38,20 @@ class Article extends CI_Controller {
 	function get_list_article($start, $limit) {
 		$result = $this->article_model->get_article_list($start, $limit);
 	 	$data_array = ""; $i = 1;
-	 	foreach($result as $row) {
-	 		$id = $row->article_id;
-        	$data_array .= "<tr><td>" . $id . "</td>";
-        	$data_array .= "<td>" . $row->title_category . "</td>";
-        	$data_array .= "<td>" . $row->title_article . "</td>";
-        	$data_array .= "<td>" . $row->status . "</td>";
-        	$data_array .= "<td>" . $row->created_date . "</td>";
-        	$data_array .= "<td>" . $row->modified_date . "</td>";
-        	$data_array .= "<td><a href='". base_url()."admin/article/detail/".$id."'>Detail</a>&nbsp;<a href='". base_url()."admin/article/update/".$id."'>Edit</a>&nbsp;<a href='". base_url() ."admin/article/delete/".$id."' onclick='return ConfirmDelete();'>Delete</a></td></tr>";
-        	$i++;
-        }
-
-        return $data_array . "</tr>";
+	 	if($result) {
+	 		foreach($result as $row) {
+		 		$id = $row->article_id;
+	        	$data_array .= "<tr><td>" . $id . "</td>";
+	        	$data_array .= "<td>" . $row->title_category . "</td>";
+	        	$data_array .= "<td>" . $row->title_article . "</td>";
+	        	$data_array .= "<td>" . $row->status . "</td>";
+	        	$data_array .= "<td>" . $row->created_date . "</td>";
+	        	$data_array .= "<td>" . $row->modified_date . "</td>";
+	        	$data_array .= "<td><a href='". base_url()."admin/article/detail/".$id."'>Detail</a>&nbsp;<a href='". base_url()."admin/article/update/".$id."'>Edit</a>&nbsp;<a href='". base_url() ."admin/article/delete/".$id."' onclick='return ConfirmDelete();'>Delete</a></td></tr>";
+	        	$i++;
+	        }
+	        return $data_array . "</tr>";	
+	 	} else return "";
 	}
 
 	function get_category_article($flag=1, $id='') {
